Introduce `use_context` to inject mini_racer
This allow devs to customize the mini_racer context, e.g, make use of
snapshots, additional options, `attach`es, etc.
