+++ NEWS.md
@@ -1,0 +1,7 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+## Upcoming
+
+Prefer injection of the custom mini_racer context
