+++ lib/humid/version.rb
@@ -1,3 +1,3 @@
-module Humid
+class Humid
   VERSION = "0.0.6".freeze
 end
